Finished basic database integration
1. Page now pulls all info (items & categories) from the database and
builds the cards from that information.
2. Colors are now saved in a separate php file for easy changing.
3. Because of this, the color part of the css has moved to the header of
the main php file.
4. Everything still seems to be working!

NOTES TO SELF
1. I may remove some of the js to save from dumping a whole database
into variables. It saved me work, but in the end, I think it's lazy
coding on my part.
2. If I keep the js, I need to un-hard-code the fill_section area. Right
now if you change one category in the database the whole thing breaks.
3. Need to write out category descriptions.
4. Would like to make categories collapsable (sp?) by player by clicking
the title of them.
5. Need to make the admin portal as well as an passwordless version for
potential users to test with.
6. Should probably clean up the main project folder and move most files
to subfolders.